Interview-Scorer/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
python-dotenv==1.0.0
langchain-core==0.2.43
langchain-groq==0.1.6
pydantic==2.4.2
typing-extensions==4.8.0
